add status errors

+++ app.js
@@ -4,7 +4,6 @@ const cors = require('cors');
 const app = express();
 const Schema = mongoose.Schema;
 
-
 const taskScheme = new Schema({
   text: String,
   isCheck: Boolean
@@ -24,27 +23,43 @@ app.get('/allTasks', (req, res) => {
 })
 
 app.post('/createTask', (req, res) => {
-  const task = new Task(req.body);
-  task.save().then(result => {
-    res.send('task created');
-  });
+  if (req.body.hasOwnProperty("text") && req.body.hasOwnProperty("isCheck")) {
+    const task = new Task(req.body);
+    task.save().then(result => {
+      res.send('task created');
+    });
+  } else {
+    res.status(500).send('not exist');
+  };
 });
 
 
 app.delete('/deleteTask', (req, res) => {
-  Task.deleteOne({ _id: req.query._id }).then((result) => {
-    Task.find().then((result) => {
-      res.send({ data: result });
+  if (req.query._id) {
+    Task.deleteOne({ _id: req.query._id }).then((result) => {
+      Task.find().then((result) => {
+        res.send({ data: result });
+      });
     });
-  });
+  } else {
+    res.status(404).send('not found');
+  }
 });
 
 app.patch('/updateTask', (req, res) => {
-  Task.updateOne({ _id: req.query._id }, req.body).then((result) => {
-    Task.find().then((result) => {
-      res.send({ data: result });
-    });
-  })
+  if (req.query._id) {
+    if (req.body.hasOwnProperty("text") && req.body.hasOwnProperty("isCheck")) {
+      Task.updateOne({ _id: req.query._id }, req.body).then((result) => {
+        Task.find().then((result) => {
+          res.send({ data: result });
+        });
+      });
+    } else {
+      res.status(500).send('not exist');
+    }
+  } else {
+    res.status(404).send('not found');
+  }
 });
 
 app.listen(8000, () => {
